backend/requirements.txt
fastapi
uvicorn[standard]
httpx
requests
python-multipart
python-dotenv
supabase
reportlab

langgraph>=0.2.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
duckduckgo-search>=6.0.0
tavily-python>=0.3.0
pydantic>=2.0.0
